change value range of port priority
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@35593 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/mstp/mstp.h
@@ -668,7 +668,7 @@ int MSTP_IN_set_cist_bridge_config(bridge_t *br, CIST_BridgeConfig *cfg);
 /* 12.8.1.4 Set MSTI Bridge Protocol Parameters */
     /* No need in special structure for single parameter Bridge Priority */
 
-int MSTP_IN_set_msti_bridge_config(tree_t *tree, __u8 bridge_priority);
+int MSTP_IN_set_msti_bridge_config(tree_t *tree, __u16 bridge_priority);
 
 /* 12.8.2.1 Read CIST Port Parameters */
 typedef struct
